fix: Minor multi-targetting fixes
