+++ src/Uno.UI/UI/Xaml/Controls/ScrollContentPresenter/IScrollContentPresenter.cs
@@ -2,22 +2,6 @@
 using System.Collections.Generic;
 using System.Text;
 using Windows.Foundation;
-#if XAMARIN_ANDROID
-using View = Android.Views.View;
-using Font = Android.Graphics.Typeface;
-#elif XAMARIN_IOS_UNIFIED
-using UIKit;
-using View = UIKit.UIView;
-using Color = UIKit.UIColor;
-using Font = UIKit.UIFont;
-#elif __MACOS__
-using AppKit;
-using View = System.Object;
-using Color = AppKit.NSColor;
-using Font = AppKit.NSFont;
-#else
-using View = System.Object;
-#endif
 
 namespace Windows.UI.Xaml.Controls
 {
@@ -30,7 +14,7 @@ namespace Windows.UI.Xaml.Controls
 	{
 		ScrollBarVisibility HorizontalScrollBarVisibility { get; set; }
 		ScrollBarVisibility VerticalScrollBarVisibility { get; set; }
-		View Content { get; set; }
+		object Content { get; set; }
 
 		void OnMinZoomFactorChanged(float newValue);
 		void OnMaxZoomFactorChanged(float newValue);
